+++ data/migrations/Version20150312203643.php
@@ -4,6 +4,7 @@ namespace Application\Migrations;
 
 use Doctrine\DBAL\Migrations\AbstractMigration,
     Doctrine\DBAL\Schema\Schema;
+use Prooph\Processing\Environment\Schema\EventStoreDoctrineSchema;
 
 /**
  * Auto-generated Migration: Please modify to your need!
@@ -12,21 +13,11 @@ class Version20150312203643 extends AbstractMigration
 {
     public function up(Schema $schema)
     {
-        $processStream = $schema->createTable('process_stream');
-
-        $processStream->addColumn('eventId', 'string', ['length' => 36]);
-        $processStream->addColumn('version', 'integer');
-        $processStream->addColumn('eventName', 'string', ['length' => 100]);
-        $processStream->addColumn('payload', 'text');
-        $processStream->addColumn('occurredOn', 'string', ['length' => 100]);
-        $processStream->addColumn('aggregate_id', 'string', ['length' => 36]);
-        $processStream->addColumn('aggregate_type', 'string', ['length' => 100]);
-        $processStream->setPrimaryKey(['eventId']);
-        $processStream->addUniqueIndex(['aggregate_id', 'aggregate_type', 'version'], 'metadata_version_uix');
+        EventStoreDoctrineSchema::createSchema($schema);
     }
 
     public function down(Schema $schema)
     {
-        $schema->dropTable('process_stream');
+        EventStoreDoctrineSchema::dropSchema($schema);
     }
 }